databay_udp_proxy/requirements.txt
aiocoap
u-msgpack-python